reduce the assumptions being made in withdrawCredits
The previous tests assumed that we can expect the amount put in from a sale is equal to the amount to withdraw. However, this is not true because other tests may be withdrawing and depositing to the address. The tests should be written in such a way that ohter tests may change the state or run each test in isolation